Downgraded java version to 1.8 for my old device + migrated Recursion I
