+++ coollime/src/main/java/coollime/recursion/atothepowerofb/Solution.java
@@ -1,0 +1,37 @@
+package coollime.recursion.atothepowerofb;
+
+public class Solution {
+
+  public long power(int a, int b) {
+    if (b == 0) {
+      return 1;
+    }
+    // Divide b by 1/2 each time and do result * result
+    long half = power(a, b / 2);
+    // We need to differentiate the situations where b is odd or even
+    if (b % 2 == 0) {
+      return half * half;
+    }
+    return half * half * a;
+  }
+
+  /**
+   * There are two simpler but more naive ways to do this problem
+   **/
+  public long naivePower(int a, int b) {
+    // Method 1: directly computation one level at a time
+    // if (b == 0) {
+    //   return 1;
+    // }
+    // return a * power(a, b - 1);
+
+    // Method 2: use a mid point (b/2) and compute the left and right halves
+    if (b == 0) {
+      return 1;
+    } else if (b == 1) {
+      return a;
+    }
+    int mid = b / 2;
+    return power(a, mid) * power(a, b - mid);
+  }
+}
