+++ docs/recursion/i/AToThePowerOfB.md
@@ -1,69 +1,40 @@
-<!----- Conversion time: 3.395 seconds.
-
-
-Using this Markdown file:
-
-1. Cut and paste this output into your source file.
-2. See the notes and action items below regarding this conversion run.
-3. Check the rendered output (headings, lists, code blocks, tables) for proper
-   formatting and use a linkchecker before you publish this page.
-
-Conversion notes:
-
-* GD2md-html version 1.0β13
-* Tue Jan 15 2019 18:28:37 GMT-0800 (PST)
-* Source doc: https://docs.google.com/open?id=1_Fp6G6RVNgm88szHAwcld4vsCdTlHLcRMX73N3gozsE
-* This document has images: check for >>>>>  gd2md-html alert:  inline image link in generated source and store images to your server.
------>
-
-
-
 # A to the Power of B
 
 [https://app.laicode.io/app/problem/13](https://app.laicode.io/app/problem/13)
 
-
 ## Description
-
 
     Evaluate a to the power of b, assuming both a and b are integers and b is non-negative.
 
 
     Examples
 
+- power(2, 0) = 1
+- power(2, 3) = 8
+- power(0, 10) = 0
+- power(-2, 5) = -32
 
+  Corner Cases
 
-*   power(2, 0) = 1
-*   power(2, 3) = 8
-*   power(0, 10) = 0
-*   power(-2, 5) = -32
-
-    Corner Cases
-
-*   What if the result is overflowed? We can assume the result will not be overflowed when we solve this problem on this online judge.
+- What if the result is overflowed? We can assume the result will not be overflowed when we solve this problem on this online judge.
 
 Medium
 
 Math
 
-
 ## Assumption
 
-
-
-*   The input should both be integers (b has to be non-negative).
-*   What if the result is overflowed? We can assume the result will not be overflowed when we solve this problem on this online judge.
-
+- The input should both be integers (b has to be non-negative).
+- What if the result is overflowed? We can assume the result will not be overflowed when we solve this problem on this online judge.
 
 ## Algorithm
-
-
 
 1.  Naive method
     1.  Base case: b == 0 ⇒ (any integer) ^ 0 == 1
     1.  Calculate the result one level at a time by calling power(a, b - 1);
-    1.  return a * power(a, b - 1);
+    1.  return a \* power(a, b - 1);
 1.  Improved method
+
     1.  Base case
         1.  b == 0 ⇒ 0
         1.  b == 1 ⇒ a because there could be odd number of subproblems
@@ -73,36 +44,30 @@ Math
 
 
     1.  return power(a, mid) * power(a, b - mid);
+
 1.  Optimized method
     1.  Base case: b == 0
     1.  Calculate the result by getting half result first, and manipulate upon the half result
-        1.  if b is even ⇒ return half * half
-        1.  if b is odd ⇒ return half * half * a
-
-
-
+        1.  if b is even ⇒ return half \* half
+        1.  if b is odd ⇒ return half _ half _ a
 
 ## Solution
 
-
 ### Method 1: naive
-
 
 #### Code
 
-
 ```java
 public class Solution {
-    public long power(int a, int b) {
-        if (b == 0) {
-            return 1;
-        }
-        return power(a, b - 1) * a;
+
+  public long power(int a, int b) {
+    if (b == 0) {
+      return 1;
     }
+    return power(a, b - 1) * a;
+  }
 }
 ```
-
-
 
 #### Complexity
 
@@ -118,10 +83,7 @@ Space:
 
 n nodes to check ⇒ O(n)
 
-
 ![naive_complexity](naive_complexity.png "image_tooltip")
-
-
 
 ### Method 2: improved → check both ways
 
@@ -129,7 +91,7 @@ The amount of computation can be lower if we divide the problem/subproblems in h
 
 For example, to compute 2^1001
 
-2^1001 == 2^500 * 2^501
+2^1001 == 2^500 \* 2^501
 
                                                                 2^1001
 
@@ -147,25 +109,22 @@ Time = O(b)
 
 Space = O(log(b))
 
-
 #### Code
-
 
 ```java
 public class Solution {
-    public long power(int a, int b) {
-        if (b == 0) {
-            return 1;
-        } else if (b == 1) {
-            return a;
-        }
-        int mid = b / 2;
-        return power(a, mid) * power(a, b - mid);
+
+  public long power(int a, int b) {
+    if (b == 0) {
+      return 1;
+    } else if (b == 1) {
+      return a;
     }
+    int mid = b / 2;
+    return power(a, mid) * power(a, b - mid);
+  }
 }
 ```
-
-
 
 #### Complexity
 
@@ -173,35 +132,29 @@ Time: total number of nodes = b ⇒ O(b)
 
 Space: log(b) levels of recursion tree that has b nodes ⇒ O(log(b))
 
-
 ![improved_complexity](improved_complexity.png "image_tooltip")
-
-
 
 ### Method 3: optimized → divide into two halves and combine
 
 We can further divide the cases into two situations: odd vs even powers. Such that the time complexity can be decreased to O(log(b))
 
-
 #### Code
-
 
 ```java
 public class Solution {
-    public long power(int a, int b) {
-        if (b == 0) {
-            return 1;
-        }
-        long half = power(a, b / 2);
-        if (b % 2 == 0) {
-            return half * half;
-        }
-        return (long) a * half * half;
+
+  public long power(int a, int b) {
+    if (b == 0) {
+      return 1;
     }
+    long half = power(a, b / 2);
+    if (b % 2 == 0) {
+      return half * half;
+    }
+    return (long) a * half * half;
+  }
 }
 ```
-
-
 
 #### Complexity
 
@@ -209,9 +162,4 @@ Time: there are only log(n) levels in the recursion tree because we divide the n
 
 Space: since the recursion tree has log(n) levels ⇒ O(log(n)).
 
-
 ![optimized_complexity](optimized_complexity.png "image_tooltip")
-
-
-
-<!-- GD2md-html version 1.0β13 -->
