+++ docs/recursion/i/README.md
@@ -1,0 +1,4 @@
+## Recursion I
+
+1.  [Fibonacci Number](FibonacciNumber.md)
+2.  [A to the Power of B](AToThePowerOfB.md)
