+++ docs/recursion/i/improved_method.png
