+++ docs/recursion/i/naive_complexity.png
