+++ docs/recursion/i/optimized_complexity.png
