--- src/C/Recursion/I/README.md
@@ -1,3 +1,0 @@
-## Recursion I
-1.  [Fibonacci Number](Easy/FibonacciNumber)
-2.  [A to the Power of B](Medium/AToThePowerOfB)
